Move synchronized to module :common-api
