+++ library/common-api/api/common-api.api
@@ -1,3 +1,12 @@
+public abstract interface annotation class io/matthewnelson/kmp/tor/common/api/ExperimentalKmpTorApi : java/lang/annotation/Annotation {
+}
+
+public abstract interface annotation class io/matthewnelson/kmp/tor/common/api/InternalKmpTorApi : java/lang/annotation/Annotation {
+}
+
+public abstract interface annotation class io/matthewnelson/kmp/tor/common/api/KmpTorDsl : java/lang/annotation/Annotation {
+}
+
 public abstract class io/matthewnelson/kmp/tor/common/api/ResourceInstaller {
 	public final field installationDir Ljava/io/File;
 	public fun <init> (Ljava/io/File;)V
@@ -17,12 +26,11 @@ public final class io/matthewnelson/kmp/tor/common/api/ResourceInstaller$Paths$T
 	public fun toString ()Ljava/lang/String;
 }
 
-public abstract interface annotation class io/matthewnelson/kmp/tor/common/api/annotation/ExperimentalKmpTorApi : java/lang/annotation/Annotation {
+public final class io/matthewnelson/kmp/tor/common/api/internal/SynchronizedCommon {
+	public static final fun synchronized (Ljava/lang/Object;Lkotlin/jvm/functions/Function0;)Ljava/lang/Object;
 }
 
-public abstract interface annotation class io/matthewnelson/kmp/tor/common/api/annotation/InternalKmpTorApi : java/lang/annotation/Annotation {
-}
-
-public abstract interface annotation class io/matthewnelson/kmp/tor/common/api/annotation/KmpTorDsl : java/lang/annotation/Annotation {
+public final class io/matthewnelson/kmp/tor/common/api/internal/SynchronizedJvm {
+	public static final fun synchronizedImpl (Ljava/lang/Object;Lkotlin/jvm/functions/Function0;)Ljava/lang/Object;
 }
