+++ library/common-api/build.gradle.kts
@@ -26,5 +26,11 @@ kmpConfiguration {
                 }
             }
         }
+
+        kotlin {
+            sourceSets.findByName("nativeMain")?.dependencies {
+                implementation(libs.kotlinx.atomicfu)
+            }
+        }
     }
 }
