+++ library/common-api/src/jvmMain/java9/module-info.java
@@ -3,5 +3,5 @@ module io.matthewnelson.kmp.tor.common.api {
     requires transitive io.matthewnelson.kmp.file;
 
     exports io.matthewnelson.kmp.tor.common.api;
-    exports io.matthewnelson.kmp.tor.common.api.annotation;
+    exports io.matthewnelson.kmp.tor.common.api.internal;
 }
