+++ library/common-core/build.gradle.kts
@@ -65,7 +65,6 @@ kmpConfiguration {
                         implementation(libs.encoding.base16)
                         implementation(libs.encoding.base64)
                         implementation(libs.kotlincrypto.hash.sha2)
-                        implementation(libs.kotlinx.atomicfu)
                     }
                 }
