+++ library/common-core/src/commonMain/kotlin/io/matthewnelson/kmp/tor/common/core/internal/-CommonPlatform.kt
@@ -18,7 +18,7 @@
 package io.matthewnelson.kmp.tor.common.core.internal
 
 import io.matthewnelson.kmp.file.File
-import io.matthewnelson.kmp.tor.common.api.annotation.InternalKmpTorApi
+import io.matthewnelson.kmp.tor.common.api.InternalKmpTorApi
 import io.matthewnelson.kmp.tor.common.core.Resource
 
 @Suppress("NOTHING_TO_INLINE")
