+++ library/common-core/src/jsTest/kotlin/io/matthewnelson/kmp/tor/common/core/JsResourceUnitTest.kt
@@ -19,7 +19,7 @@ import io.matthewnelson.kmp.file.SysTempDir
 import io.matthewnelson.kmp.file.name
 import io.matthewnelson.kmp.file.readUtf8
 import io.matthewnelson.kmp.file.resolve
-import io.matthewnelson.kmp.tor.common.api.annotation.InternalKmpTorApi
+import io.matthewnelson.kmp.tor.common.api.InternalKmpTorApi
 import kotlin.test.Test
 import kotlin.test.assertEquals
