+++ library/common-core/src/jsTest/kotlin/io/matthewnelson/kmp/tor/common/core/OSInfoUnitTest.kt
@@ -16,7 +16,7 @@
 package io.matthewnelson.kmp.tor.common.core
 
 import io.matthewnelson.kmp.file.resolve
-import io.matthewnelson.kmp.tor.common.api.annotation.InternalKmpTorApi
+import io.matthewnelson.kmp.tor.common.api.InternalKmpTorApi
 import kotlin.test.Test
 import kotlin.test.assertTrue
