+++ library/common-core/src/jvmMain/kotlin/io/matthewnelson/kmp/tor/common/core/OSInfo.kt
@@ -21,7 +21,7 @@ package io.matthewnelson.kmp.tor.common.core
 import io.matthewnelson.kmp.file.ANDROID
 import io.matthewnelson.kmp.file.File
 import io.matthewnelson.kmp.file.toFile
-import io.matthewnelson.kmp.tor.common.api.annotation.InternalKmpTorApi
+import io.matthewnelson.kmp.tor.common.api.InternalKmpTorApi
 import io.matthewnelson.kmp.tor.common.core.internal.ARCH_MAP
 import io.matthewnelson.kmp.tor.common.core.internal.PATH_MAP_FILES
 import io.matthewnelson.kmp.tor.common.core.internal.PATH_OS_RELEASE
