+++ library/common-core/src/jvmMain/kotlin/io/matthewnelson/kmp/tor/common/core/internal/-JvmPlatform.kt
@@ -17,7 +17,7 @@ package io.matthewnelson.kmp.tor.common.core.internal
 
 import io.matthewnelson.kmp.file.File
 import io.matthewnelson.kmp.file.IOException
-import io.matthewnelson.kmp.tor.common.api.annotation.InternalKmpTorApi
+import io.matthewnelson.kmp.tor.common.api.InternalKmpTorApi
 import io.matthewnelson.kmp.tor.common.core.Resource
 import java.util.zip.GZIPInputStream
