+++ library/common-core/src/nativeMain/kotlin/io/matthewnelson/kmp/tor/common/core/NativeResource.kt
@@ -19,7 +19,7 @@ import io.matthewnelson.encoding.base16.Base16
 import io.matthewnelson.encoding.base64.Base64
 import io.matthewnelson.encoding.core.Decoder.Companion.decodeToByteArray
 import io.matthewnelson.encoding.core.Encoder.Companion.encodeToString
-import io.matthewnelson.kmp.tor.common.api.annotation.InternalKmpTorApi
+import io.matthewnelson.kmp.tor.common.api.InternalKmpTorApi
 import org.kotlincrypto.hash.sha2.SHA256
 
 /**
