+++ library/common-core/src/nativeTest/kotlin/io/matthewnelson/kmp/tor/common/core/NativeResourceUnitTest.kt
@@ -16,7 +16,7 @@
 package io.matthewnelson.kmp.tor.common.core
 
 import io.matthewnelson.kmp.file.*
-import io.matthewnelson.kmp.tor.common.api.annotation.InternalKmpTorApi
+import io.matthewnelson.kmp.tor.common.api.InternalKmpTorApi
 import kotlin.test.*
 
 @OptIn(InternalKmpTorApi::class)
