+++ library/common-core/src/nonNativeMain/kotlin/io/matthewnelson/kmp/tor/common/core/OSArch.kt
@@ -17,7 +17,7 @@
 
 package io.matthewnelson.kmp.tor.common.core
 
-import io.matthewnelson.kmp.tor.common.api.annotation.InternalKmpTorApi
+import io.matthewnelson.kmp.tor.common.api.InternalKmpTorApi
 import kotlin.jvm.JvmField
 
 /**
