+++ library/common-core/src/nonNativeMain/kotlin/io/matthewnelson/kmp/tor/common/core/internal/-NonNativePlatform.kt
@@ -16,7 +16,7 @@
 package io.matthewnelson.kmp.tor.common.core.internal
 
 import io.matthewnelson.immutable.collections.immutableMapOf
-import io.matthewnelson.kmp.tor.common.api.annotation.InternalKmpTorApi
+import io.matthewnelson.kmp.tor.common.api.InternalKmpTorApi
 import io.matthewnelson.kmp.tor.common.core.OSArch
 import kotlin.jvm.JvmSynthetic
